+++ src/org/launchcode/java/demos/lsn5unittesting/main/Car.java
@@ -85,4 +85,8 @@ public class Car {
         this.odometer += milesAbleToTravel;
     }
 
+    public void addGas(double gas) {
+        this.setGasTankLevel(gas + this.getGasTankLevel());
+    }
+
 }
