+++ src/org/launchcode/java/demos/lsn5unittesting/test/CarTest.java
@@ -16,20 +16,36 @@ public class CarTest {
 
     //TODO: constructor sets gasTankLevel properly
     Car test_car;
+
     @Before
     public void createCarObject(){
-        Car test_car = new Car("Toyota", "Prius", 10, 50);
+        test_car = new Car("Toyota", "Prius", 10, 50);
+        test_car.drive(50);
+        test_car.drive(500);
     }
     @Test
     public void testInitialGasTank() {
-        Car test_car;
+
         assertEquals(10, test_car.getGasTankLevel(), .001);
-
-
 
     }
     //TODO: gasTankLevel is accurate after driving within tank range
+
+    @Test
+    public void testGasTankAfterDriving() {
+        assertEquals(9, test_car.getGasTankLevel(), .001);
+    }
+
     //TODO: gasTankLevel is accurate after attempting to drive past tank range
+    @Test
+    public void testGasTankAfterExceedingTankRange() {
+        assertEquals(0, test_car.getGasTankLevel(), .001);
+    }
     //TODO: can't have more gas than tank size, expect an exception
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testGasTankOverfillException() {
+
+    }
 
 }
